+++ lib/src/package_context.dart
@@ -3,13 +3,15 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'dart:convert';
+import 'dart:io';
 
 import 'package:collection/collection.dart';
-import 'package:pana/src/dartdoc/dartdoc_options.dart';
+import 'package:path/path.dart' as p;
 import 'package:pub_semver/pub_semver.dart';
 
 import 'code_problem.dart';
 import 'dartdoc/dartdoc.dart';
+import 'dartdoc/dartdoc_options.dart';
 import 'download_utils.dart';
 import 'internal_model.dart';
 import 'license.dart';
@@ -24,7 +26,7 @@ import 'repository/check_repository.dart';
 import 'screenshots.dart';
 import 'sdk_env.dart';
 import 'tag/license_tags.dart';
-import 'utils.dart' show listFocusDirs;
+import 'utils.dart' show listFocusDirs, PanaProcessResult;
 
 /// Shared (intermediate) results between different packages or versions.
 /// External systems that may be independent of the archive content may be
@@ -226,46 +228,69 @@ class PackageContext {
     return LicenseTags.fromLicenses(await licenses);
   }();
 
-  late final dartdocReportSubsection = () async {
+  late final Future<DartdocResult?> dartdocResult = () async {
     final dartdocOutputDir = options.dartdocOutputDir;
     if (dartdocOutputDir == null) {
       return null;
     }
-    DartdocResult? dartdocResult;
-    String? errorReason;
     if (await resolveDependencies()) {
       final timeout = options.dartdocTimeout ?? const Duration(minutes: 5);
       await normalizeDartdocOptionsYaml(packageDir);
       try {
-        dartdocResult = await toolEnvironment.dartdoc(
+        final pr = await toolEnvironment.dartdoc(
           packageDir,
           dartdocOutputDir,
           validateLinks: false,
           timeout: timeout,
           usesFlutter: usesFlutter,
         );
-        if (dartdocResult.wasTimeout) {
-          errorReason = '`dartdoc` could not complete in $timeout.';
+        if (pr.exitCode == 15) {
+          return DartdocResult(
+              errorReason: '`dartdoc` could not complete in $timeout.');
         }
+        if (pr.exitCode != 0) {
+          return DartdocResult(errorReason: pr.asTrimmedOutput);
+        }
+
+        final hasIndexHtml =
+            await File(p.join(dartdocOutputDir, 'index.html')).exists();
+        final hasIndexJson =
+            await File(p.join(dartdocOutputDir, 'index.json')).exists();
+        return DartdocResult(
+            processResult: pr, hasOutputFiles: hasIndexHtml && hasIndexJson);
       } catch (e, st) {
-        errorReason = 'Could not run `dartdoc`: $e';
         log.severe('Could not run dartdoc.', e, st);
+        return DartdocResult(errorReason: 'Could not run `dartdoc`: $e');
       }
     } else {
-      return dartdocFailedSubsection(
-          'Dependency resultion failed, unable to run `dartdoc`.');
+      return DartdocResult(
+          errorReason: 'Dependency resultion failed, unable to run `dartdoc`.');
     }
-
-    if (dartdocResult == null) {
-      return dartdocFailedSubsection(errorReason ?? 'Could not run `dartdoc`.');
-    }
-
-    if (!dartdocResult.wasSuccessful) {
-      return dartdocFailedSubsection(
-          dartdocResult.processResult.asTrimmedOutput);
-    }
-
-    final pubData = await generateAndSavePubDataJson(dartdocOutputDir);
-    return await createDocumentationCoverageSection(pubData);
   }();
+
+  late final dartdocPubData = () async {
+    final dr = await dartdocResult;
+    if (dr == null || !dr.wasSuccessful) {
+      return null;
+    }
+    return await generateAndSavePubDataJson(options.dartdocOutputDir!);
+  }();
+}
+
+class DartdocResult {
+  final PanaProcessResult? processResult;
+  final String? errorReason;
+  final bool hasOutputFiles;
+
+  DartdocResult({
+    this.processResult,
+    this.errorReason,
+    this.hasOutputFiles = false,
+  });
+
+  bool get wasSuccessful =>
+      processResult != null &&
+      processResult!.exitCode == 0 &&
+      errorReason == null &&
+      hasOutputFiles;
 }